requirements.txt
Django>=4.0
psycopg2>=2.8
djangorestframework>=3.13.1
markdown>=3.3.6
django-filter>=21.1